+++ ouds_core/lib/components/button/internal/ouds_button_border_modifier.dart
@@ -10,6 +10,9 @@
 // Software description: Flutter library of reusable graphical components
 //
 
+/// @nodoc
+library;
+
 import 'package:flutter/material.dart';
 import 'package:ouds_core/components/button/internal/ouds_button_control_state.dart';
 import 'package:ouds_core/components/button/internal/ouds_button_loading_modifier.dart';
@@ -46,7 +49,7 @@ class OudsButtonBorderModifier {
     final onColoredSurface = OudsTheme.isOnColoredSurfaceOf(context);
     switch (hierarchy) {
       case OudsButtonHierarchy.strong:
-        return onColoredSurface ? BorderSide(color: theme.componentsTokens(context).buttonMono.colorBorderStrongEnabled, width: theme.componentsTokens(context).button.borderWidthDefault) : BorderSide.none;
+        return BorderSide.none;
       case OudsButtonHierarchy.brand:
         return BorderSide.none;
       case OudsButtonHierarchy.minimal:
@@ -64,7 +67,7 @@ class OudsButtonBorderModifier {
     final onColoredSurface = OudsTheme.isOnColoredSurfaceOf(context);
     switch (hierarchy) {
       case OudsButtonHierarchy.strong:
-        return onColoredSurface ? BorderSide(color: theme.componentsTokens(context).buttonMono.colorBorderStrongHover, width: theme.componentsTokens(context).button.borderWidthDefaultInteraction) : BorderSide.none;
+        return BorderSide.none;
       case OudsButtonHierarchy.brand:
         return BorderSide.none;
       case OudsButtonHierarchy.minimal:
@@ -82,7 +85,7 @@ class OudsButtonBorderModifier {
     final onColoredSurface = OudsTheme.isOnColoredSurfaceOf(context);
     switch (hierarchy) {
       case OudsButtonHierarchy.strong:
-        return onColoredSurface ? BorderSide(color: theme.componentsTokens(context).buttonMono.colorBorderStrongPressed, width: theme.componentsTokens(context).button.borderWidthDefaultInteraction) : BorderSide.none;
+        return BorderSide.none;
       case OudsButtonHierarchy.brand:
         return BorderSide.none;
       case OudsButtonHierarchy.minimal:
@@ -101,7 +104,7 @@ class OudsButtonBorderModifier {
     final onColoredSurface = OudsTheme.isOnColoredSurfaceOf(context);
     switch (hierarchy) {
       case OudsButtonHierarchy.strong:
-        return onColoredSurface ? BorderSide(color: theme.componentsTokens(context).buttonMono.colorBorderStrongDisabled, width: theme.componentsTokens(context).button.borderWidthDefault) : BorderSide.none;
+        return BorderSide.none;
       case OudsButtonHierarchy.brand:
         return BorderSide.none;
       case OudsButtonHierarchy.minimal: