+++ ouds_core/lib/components/button/ouds_button.dart
@@ -9,6 +9,9 @@
 // Software description: Flutter library of reusable graphical components
 //
 
+/// OudsButton
+library;
+
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter_svg/flutter_svg.dart';
@@ -201,6 +204,7 @@ class _OudsButtonState extends State<OudsButton> {
       case OudsButtonControlState.loading:
         return Semantics(
           label: OudsLocalizations.of(context)?.core_button_loading_a11y,
+          enabled: false,
           button: true,
           child: ExcludeSemantics(
             child: OutlinedButton(
@@ -285,6 +289,7 @@ class _OudsButtonState extends State<OudsButton> {
       case OudsButtonControlState.loading:
         return Semantics(
           label: OudsLocalizations.of(context)?.core_button_loading_a11y,
+          enabled: false,
           button: true,
           child: IconButton(
             onPressed: null,
@@ -333,6 +338,7 @@ class _OudsButtonState extends State<OudsButton> {
       case OudsButtonControlState.loading:
         return Semantics(
           label: OudsLocalizations.of(context)?.core_button_loading_a11y,
+          enabled: false,
           button: true,
           child: ExcludeSemantics(
             child: OutlinedButton(