+++ ouds_core/lib/components/control/internal/controller/ouds_interaction_state_controller.dart
@@ -8,6 +8,8 @@
 //
 // Software description: Flutter library of reusable graphical components
 //
+/// @nodoc
+library;
 
 import 'package:flutter/material.dart';