+++ ouds_core/lib/components/control/internal/modifier/ouds_control_border_modifier.dart
@@ -9,6 +9,9 @@
 // Software description: Flutter library of reusable graphical components
 //
 
+/// @nodoc
+library;
+
 import 'package:flutter/widgets.dart';
 import 'package:ouds_core/components/control/internal/ouds_control_state.dart';
 import 'package:ouds_theme_contract/ouds_theme.dart';
@@ -20,7 +23,7 @@ class OudsControlBorderModifier {
   OudsControlBorderModifier(this.context);
 
   /// Gets the border color based on the indicator state and error status.
-  Color getBorderColor(OudsControlState state, bool error, bool selected) {
+  Color getBorderColor(OudsControlState state, bool error, bool selected, bool _isHighContrast) {
     final colorScheme = OudsTheme.of(context).colorScheme;
     if (error) {
       // Error
@@ -44,7 +47,7 @@ class OudsControlBorderModifier {
         case OudsControlState.enabled:
           if (selected) {
             // In order to reach the a11y AAA level, the selected checkbox is black
-            return (MediaQuery.highContrastOf(context)) ? colorScheme(context).contentDefault : colorScheme(context).actionSelected;
+            return _isHighContrast ? colorScheme(context).contentDefault : colorScheme(context).actionSelected;
           } else {
             return colorScheme(context).actionEnabled;
           }
@@ -54,7 +57,7 @@ class OudsControlBorderModifier {
           return colorScheme(context).actionHover;
         case OudsControlState.pressed:
           // In order to reach the a11y AAA level, the pressed checkbox is black
-          return (MediaQuery.highContrastOf(context)) ? colorScheme(context).contentDefault : colorScheme(context).actionPressed;
+          return _isHighContrast ? colorScheme(context).contentDefault : colorScheme(context).actionPressed;
         case OudsControlState.focused:
           return colorScheme(context).actionFocus;
         case OudsControlState.readOnly:
@@ -82,7 +85,7 @@ class OudsControlBorderModifier {
   }
 
   /// Gets the border radius based on the checkbox.
-  double getBorderRadius(token) {
+  double getBorderRadius(dynamic token) {
     return token.borderRadius;
   }
 }