+++ ouds_core/lib/components/control/ouds_control_item.dart
@@ -9,9 +9,13 @@
 // Software description: Flutter library of reusable graphical components
 //
 
+/// @nodoc
+library;
+
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+import 'package:ouds_accessibility_plugin/ouds_accessibility_plugin.dart';
 import 'package:ouds_core/components/control/internal/controller/ouds_interaction_state_controller.dart';
 import 'package:ouds_core/components/control/internal/interaction/ouds_inherited_interaction_model.dart';
 import 'package:ouds_core/components/control/internal/modifier/ouds_control_background_modifier.dart';
@@ -90,6 +94,7 @@ class OudsControlItem extends StatefulWidget {
 class OudsControlItemState extends State<OudsControlItem> {
   // Create an instance of the state controller to manage interaction changes
   final OudsInteractionStateController interactionState = OudsInteractionStateController();
+  bool _isHighContrast = false;
 
   @override
   void initState() {
@@ -111,6 +116,16 @@ class OudsControlItemState extends State<OudsControlItem> {
   }
 
   @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    OudsAccessibilityPlugin.isHighContrastEnabled(context).then((value) {
+      setState(() {
+        _isHighContrast = value;
+      });
+    });
+  }
+
+  @override
   Widget build(BuildContext context) {
     final controlItemStateDeterminer = OudsControlStateDeterminer(
       enabled: widget.onTap != null,
@@ -127,7 +142,7 @@ class OudsControlItemState extends State<OudsControlItem> {
       state: interactionState,
       child: Padding(
         padding: EdgeInsetsDirectional.symmetric(
-          horizontal: OudsTheme.of(context).componentsTokens(context).controlItem.spaceInset,
+          horizontal: OudsTheme.of(context).componentsTokens(context).controlItem.spacePaddingBlock,
         ),
         child: Stack(
           children: [
@@ -147,14 +162,14 @@ class OudsControlItemState extends State<OudsControlItem> {
                   child: InkWell(
                     onTap: !widget.readOnly
                         ? () {
-                            interactionState.setPressed(true);
-                            // Added to improve visual rendering fluidity by allowing Flutter
-                            // to complete the current frame before executing the state change logic.
-                            SchedulerBinding.instance.addPostFrameCallback((_) {
-                              widget.onTap?.call();
-                              interactionState.setPressed(false);
-                            });
-                          }
+                      interactionState.setPressed(true);
+                      // Added to improve visual rendering fluidity by allowing Flutter
+                      // to complete the current frame before executing the state change logic.
+                      SchedulerBinding.instance.addPostFrameCallback((_) {
+                        widget.onTap?.call();
+                        interactionState.setPressed(false);
+                      });
+                    }
                         : null,
                     onHighlightChanged: widget.onTap != null ? interactionState.setPressed : null,
                     onHover: interactionState.setHovered,
@@ -163,7 +178,7 @@ class OudsControlItemState extends State<OudsControlItem> {
                     splashColor: Colors.transparent,
                     child: Padding(
                       padding: EdgeInsetsDirectional.all(
-                        OudsTheme.of(context).componentsTokens(context).controlItem.spaceInset,
+                        OudsTheme.of(context).componentsTokens(context).controlItem.spacePaddingBlock,
                       ),
                       child: IntrinsicHeight(
                         child: Row(
@@ -186,6 +201,7 @@ class OudsControlItemState extends State<OudsControlItem> {
                           controlItemState,
                           widget.error,
                           widget.selected,
+                          _isHighContrast,
                         ),
                         width: OudsTheme.of(context).borderTokens.widthThin,
                       ),