+++ ouds_core/lib/components/switch/ouds_switch.dart
@@ -9,9 +9,13 @@
 // Software description: Flutter library of reusable graphical components
 //
 
+/// OudsSwitch
+library;
+
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+import 'package:ouds_accessibility_plugin/ouds_accessibility_plugin.dart';
 import 'package:ouds_core/components/control/internal/interaction/ouds_inherited_interaction_model.dart';
 import 'package:ouds_core/components/control/internal/modifier/ouds_control_tick_modifier.dart';
 import 'package:ouds_core/components/control/internal/ouds_control_state.dart';
@@ -66,7 +70,17 @@ class _OudsSwitchState extends State<OudsSwitch> {
   bool _isHovered = false;
   bool _isPressed = false;
   bool _isFocused = false;
+  bool _isHighContrast = false;
 
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    OudsAccessibilityPlugin.isHighContrastEnabled(context).then((value) {
+      setState(() {
+        _isHighContrast = value;
+      });
+    });
+  }
   @override
   Widget build(BuildContext context) {
     final interactionModelHover = OudsInheritedInteractionModel.of(context, InteractionAspect.hover);
@@ -134,7 +148,10 @@ class _OudsSwitchState extends State<OudsSwitch> {
                     minWidth: switchButton.sizeMinWidth,
                     maxHeight: switchButton.sizeMaxHeight,
                   ),
-                  decoration: BoxDecoration(borderRadius: BorderRadius.circular(switchButton.borderRadiusTrack), color: switchTickModifier.getTickSwitchColor(switchState, widget.value)),
+                  decoration: BoxDecoration(
+                      borderRadius: BorderRadius.circular(switchButton.borderRadiusTrack),
+                      color: switchTickModifier.getTickSwitchColor(switchState, widget.value,_isHighContrast)
+                  ),
                   child: _buildCursorIndicator(context, switchState, isPressed, isHovered),
                 ),
               ),
@@ -170,7 +187,7 @@ class _OudsSwitchState extends State<OudsSwitch> {
             borderRadius: BorderRadius.circular(switchButton.borderRadiusCursor),
             boxShadow: [
               BoxShadow(
-                color: Colors.black.withOpacity(0.30),
+                color: Colors.black.withAlpha((0.30 * 255).round()),
                 offset: Offset(0, 1),
                 blurRadius: 2,
                 spreadRadius: 0,