+++ ouds_core/lib/modules/module-more/more_screen.dart
@@ -11,6 +11,10 @@
  */
 
 /// @nodoc
+
+/// @nodoc
+library;
+
 import 'package:flutter/material.dart';
 
 class MoreScreen extends StatelessWidget {